backend/pyproject.toml
[project]
name = "voxora-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlmodel",
    "sqlalchemy",
    "pydantic-settings",
    "python-multipart",
    "httpx",
    "pytest",
    "pytest-anyio",
    "aiofiles>=25.1.0",
    "faster-whisper>=1.0.0",
    "prometheus-client>=0.24.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
